Initial Commit
(Split linting to separate versions so easier to diff)

This is my version of the CSM template. It has one minor correction to sanity.
It also modifies paths to be "root" which assumes the user adds this directory
to their LaTeX path (e.g. using TEXINPUTS for project-specific libraries).
	export TEXINPUTS=$(pwd)/path/to/here//:
	(xelatex ...)

Root Paths for required packages